# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Reservas de Club Deportivo
  description: |
    API REST para gestionar canchas, socios y reservas del Club Deportivo Encuentro.

    El contrato sigue el enunciado del trabajo práctico. La URL base puede variar
    según el despliegue (por ejemplo, `/api` o `/v1`).
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor local de desarrollo
...
paths:
...
  /bloqueos:
    get:
      summary: Listar bloqueos de mantenimiento
      description: Extensión opcional. Listado paginado de bloqueos.
      tags:
        - Extensiones opcionales
      parameters:
        - $ref: '#/components/parameters/IdCanchaFilter'
        - $ref: '#/components/parameters/FechaBloqueoFilter'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/GetBloqueosList200'
        '204':
          $ref: '#/components/responses/NoContent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Crear un bloqueo de mantenimiento
      description: |
        Extensión opcional. Crea un bloqueo que impide reservar una cancha en
        el intervalo indicado.
      tags:
        - Extensiones opcionales
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BloqueoCreate'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /bloqueos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    delete:
      summary: Eliminar un bloqueo de mantenimiento
      description: Extensión opcional. Quita la restricción del bloqueo.
      tags:
        - Extensiones opcionales
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Bloqueo:
      type: object
      properties:
        id:
          type: integer
          description: Identificador único del bloqueo
          example: 1
        id_cancha:
          type: integer
          description: Identificador de la cancha bloqueada
          example: 2
        fecha:
          type: string
          format: date
          description: Día del bloqueo
          example: "2026-10-20"
        hora_inicio:
          type: string
          description: Hora de inicio del bloqueo
          pattern: '^([01]\d|2[0-3]):00:00$'
          example: "08:00:00"
        hora_fin:
          type: string
          description: Hora de fin del bloqueo
          pattern: '^([01]\d|2[0-3]):00:00$'
          example: "12:00:00"
        motivo:
          type: string
          description: Motivo del bloqueo
          example: "Mantenimiento de iluminación"
    BloqueoCreate:
      type: object
      required:
        - id_cancha
        - fecha
        - hora_inicio
        - hora_fin
        - motivo
      properties:
        id_cancha:
          type: integer
          description: Identificador de la cancha a bloquear
          example: 2
        fecha:
          type: string
          format: date
          description: Día del bloqueo
          example: "2026-10-20"
        hora_inicio:
          type: string
          description: Hora de inicio del bloqueo
          pattern: '^([01]\d|2[0-3]):00:00$'
          example: "08:00:00"
        hora_fin:
          type: string
          description: Hora de fin del bloqueo
          pattern: '^([01]\d|2[0-3]):00:00$'
          example: "12:00:00"
        motivo:
          type: string
          description: Motivo del bloqueo
          example: "Mantenimiento de iluminación"
...
    Links:
      type: object
      description: Enlaces HATEOAS para navegar entre páginas
      properties:
        _first:
          $ref: '#/components/schemas/Link'
        _prev:
          $ref: '#/components/schemas/Link'
        _next:
          $ref: '#/components/schemas/Link'
        _last:
          $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        href:
          type: string
          description: URI para navegar a la página indicada
          example: "http://localhost:5000/canchas?_offset=0&_limit=10"
    Error:
      type: object
      properties:
        code:
          type: string
          description: Código alfanumérico del error
          example: "ERROR_VALIDACION"
        message:
          type: string
          description: Resumen breve del error
          example: "El cuerpo de la solicitud es inválido"
        level:
          type: string
          description: Nivel del error
          enum:
            - info
            - warning
            - error
          example: "error"
        description:
          type: string
          description: Descripción detallada del error
          example: "El campo 'precio_hora' debe ser un entero mayor a cero"
    Errores:
      type: object
      properties:
        errors:
          type: array
          description: Listado de errores
          items:
            $ref: '#/components/schemas/Error'
...
    BloqueosListResponse:
      type: object
      properties:
        bloqueos:
          type: array
          items:
            $ref: '#/components/schemas/Bloqueo'
        _links:
          $ref: '#/components/schemas/Links'
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Identificador entero positivo del recurso
      schema:
        type: integer
      example: 1
    Limit:
      name: _limit
      in: query
      description: Cantidad máxima de registros por página
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
    Offset:
      name: _offset
      in: query
      description: Desplazamiento desde el primer registro
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
...
    IdCanchaFilter:
      name: id_cancha
      in: query
      description: Filtrar por identificador de cancha
      schema:
        type: integer
      example: 2
...
    FechaBloqueoFilter:
      name: fecha
      in: query
      description: Filtrar bloqueos por día
      schema:
        type: string
        format: date
      example: "2026-10-20"
  responses:
...
    GetBloqueosList200:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BloqueosListResponse'
    Created:
      description: Creado
...
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    Conflict:
      description: Conflicto de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'
    NoContent:
      description: Sin contenido
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errores'